requirements.txt
asgiref==3.7.2
boolean.py==4.0
CacheControl==0.13.1
certifi==2023.7.22
charset-normalizer==3.3.1
cyclonedx-python-lib==4.2.3
defusedxml==0.7.1
Django==4.2.8
django-auto-logout==0.5.1
django-axes==6.0.1
django-polymorphic==3.1.0
django-recaptcha==3.0.0
et-xmlfile==1.1.0
filelock==3.12.4
html5lib==1.1
idna==3.4
license-expression==30.1.1
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.0.7
mysql-connector-python==8.1.0
mysqlclient==2.2.0
openpyxl==3.1.2
packageurl-python==0.11.2
packaging==23.2
Pillow==10.1.0
pip-api==0.0.30
pip-requirements-parser==32.0.1
pip_audit==2.6.1
protobuf==4.21.12
py-serializable==0.11.1
Pygments==2.16.1
pyparsing==3.1.1
requests==2.31.0
rich==13.6.0
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1